+++ Editor/CoverageWindow/IncludedAssembliesTreeView.cs
@@ -46,40 +46,26 @@ namespace UnityEditor.TestTools.CodeCoverage
 
             if (developerMode)
             {
-                string allAssemblyFiltersString = AssemblyFiltering.GetAllProjectAssembliesString() + ",unityeditor*,unityengine*,unity.*";
-                string[] allAssemblyFilters = allAssemblyFiltersString.Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries);
-                
-                Regex[] allAssemblies = allAssemblyFilters
-                .Select(f => AssemblyFiltering.CreateFilterRegex(f))
-                .ToArray();
-
-                System.Reflection.Assembly[] assemblies = AppDomain.CurrentDomain.GetAssemblies();
-                Array.Sort(assemblies, (x, y) => String.Compare(x.GetName().Name, y.GetName().Name));
-
+                System.Reflection.Assembly[] assemblies = AssemblyFiltering.GetAllProjectAssembliesInternal();               
                 int assembliesLength = assemblies.Length;
 
                 GUIContent textContent = new GUIContent();
                 for (int i = 0; i < assembliesLength; ++i)
                 {
                     System.Reflection.Assembly assembly = assemblies[i];
+                    bool enabled = includeAssemblies.Any(f => f.IsMatch(assembly.GetName().Name.ToLowerInvariant()));
+                    root.AddChild(new AssembliesTreeViewItem() { id = i + 1, displayName = assembly.GetName().Name, Enabled = enabled });
 
-                    if (allAssemblies.Any(f => f.IsMatch(assembly.GetName().Name.ToLowerInvariant())))
-                    {
-                        bool enabled = includeAssemblies.Any(f => f.IsMatch(assembly.GetName().Name.ToLowerInvariant()));
-                        root.AddChild(new AssembliesTreeViewItem() { id = i + 1, displayName = assembly.GetName().Name, Enabled = enabled });
-
-                        textContent.text = assembly.GetName().Name;
-                        float itemWidth = TreeView.DefaultStyles.label.CalcSize(textContent).x + kCheckBoxWidth;
-                        if (Width < itemWidth)
-                            Width = itemWidth;
-                    }
+                    textContent.text = assembly.GetName().Name;
+                    float itemWidth = TreeView.DefaultStyles.label.CalcSize(textContent).x + kCheckBoxWidth;
+                    if (Width < itemWidth)
+                        Width = itemWidth;
+                    
                 }
             }
             else
             {
-                Assembly[] assemblies = CompilationPipeline.GetAssemblies();
-                Array.Sort(assemblies, (x, y) => String.Compare(x.name, y.name));
-
+                Assembly[] assemblies = AssemblyFiltering.GetAllProjectAssemblies();
                 int assembliesLength = assemblies.Length;
 
                 GUIContent textContent = new GUIContent();